jjep

+++ Ohjelmisto1Tehtävä3.1.py
@@ -1,3 +1,3 @@
-kpituus = int(input("Anna kuhan pituus sentteinä: "))
+kpituus = float(input("Anna kuhan pituus sentteinä: "))
 if kpituus <37:
     print(f"Laske kuha takaisin järveen, sen pitää kasvaa vielä {37-kpituus} senttiä ")
